# Interpolation API: switch interpolations to BSplineKit