--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -9,8 +9,8 @@ jobs:
     strategy:
       matrix:
         version:
-          - '1.5'
           - '1.6'
+          - '1.7'
           - 'nightly'
         os:
           - ubuntu-latest