--- a/Project.toml
+++ b/Project.toml
@@ -4,14 +4,13 @@ authors = ["Alec Loudenback <alecloudenback@gmail.com> and contributors"]
 version = "0.10.0"
 
 [deps]
+BSplineKit = "093aae92-e908-43d7-9660-e50ee39d5a0a"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 
 [compat]
 ForwardDiff = "^0.10"
-Interpolations = "^0.12.4, 0.13"
 UnicodePlots = "^2"
 julia = "^1.5"