--- a/Project.toml
+++ b/Project.toml
@@ -12,7 +12,7 @@ UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 [compat]
 ForwardDiff = "^0.10"
 UnicodePlots = "^2"
-julia = "^1.5"
+julia = "^1.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"