--- a/src/utils.jl
+++ b/src/utils.jl
@@ -37,23 +37,24 @@ function bootstrap(rates, maturities, settlement_frequency, interpolation::Inter
     return _bootstrap_choose_interp(rates, maturities, settlement_frequency, interpolation)
 end
 
+# the fall-back if user provides own interpolation function
+function bootstrap(rates, maturities, settlement_frequency, interpolation)
+    return _bootstrap_inner(rates, maturities, settlement_frequency, interpolation)
+end
+
 # dispatch on the user-exposed InterpolationKind to the right 
 # internally named interpolation function
 function _bootstrap_choose_interp(rates, maturities, settlement_frequency, i::CubicSpline)
-    return bootstrap_inner(rates, maturities, settlement_frequency, cubic_interp)
+    return _bootstrap_inner(rates, maturities, settlement_frequency, cubic_interp)
 end
 
 function _bootstrap_choose_interp(rates, maturities, settlement_frequency, i::LinearSpline)
-    return bootstrap_inner(rates, maturities, settlement_frequency, linear_interp)
-end
-
-# the fall-back if user provides own interpolation function
-function _bootstrap_choose_interp(rates, maturities, settlement_frequency, i_other)
-    return bootstrap_inner(rates, maturities, settlement_frequency, i_other)
+    return _bootstrap_inner(rates, maturities, settlement_frequency, linear_interp)
 end
 
 
-function bootstrap_inner(rates, maturities, settlement_frequency, interpolation_function)
+
+function _bootstrap_inner(rates, maturities, settlement_frequency, interpolation_function)
     discount_vec = zeros(length(rates)) # construct a placeholder discount vector matching maturities
     # we have to take the first rate as the starting point
     discount_vec[1] = discount(Constant(rates[1]), maturities[1])